VisualStudioの警告を修正

+++ core/common/servent.cpp
@@ -2720,7 +2720,7 @@ void Servent::sendRawMetaChannel(int interval)
 									url.convertTo(String::T_META);
 
 									snprintf(tmp, _countof(tmp),"StreamTitle='%s';StreamUrl='%s';\0",title.cstr(),url.cstr());
-									int len = ((strlen(tmp) + 15+1) / 16);
+									int len = ((static_cast<int>(strlen(tmp)) + 15+1) / 16);
 									sock->writeChar(len);
 									sock->write(tmp,len*16);
